requirements.txt
discord.py==2.3.2
google-generativeai==0.4.1
openai==1.30.1
requests==2.31.0
Flask==3.0.2